Bring over HFS+ fixes.

+++ fs/src/fs/org/jnode/fs/hfsplus/HfsPlusForkData.java
@@ -39,11 +39,11 @@ public class HfsPlusForkData {
      */
     private long totalSize;
     /** */
-    private int clumpSize;
+    private long clumpSize;
     /**
      * The total of allocation blocks use by the extents in the fork.
      */
-    private int totalBlock;
+    private long totalBlock;
     /**
      * The first eight extent descriptors for the fork.
      */
@@ -77,8 +77,8 @@ public class HfsPlusForkData {
         byte[] data = new byte[FORK_DATA_LENGTH];
         System.arraycopy(src, offset, data, 0, FORK_DATA_LENGTH);
         totalSize = BigEndian.getInt64(data, 0);
-        clumpSize = BigEndian.getInt32(data, 8);
-        totalBlock = BigEndian.getInt32(data, 12);
+        clumpSize = BigEndian.getUInt32(data, 8);
+        totalBlock = BigEndian.getUInt32(data, 12);
         extents = new ExtentDescriptor[8];
         for (int i = 0; i < 8; i++) {
             extents[i] =
@@ -89,7 +89,6 @@ public class HfsPlusForkData {
 
     /**
      * Create a new empty data-fork data.
-     * Create a new empty fork data object.
      *
      * @param totalSize
      * @param clumpSize
@@ -110,8 +109,8 @@ public class HfsPlusForkData {
     public byte[] write(byte[] dest, int destOffSet) {
         byte[] data = new byte[FORK_DATA_LENGTH];
         BigEndian.setInt64(data, 0, totalSize);
-        BigEndian.setInt32(data, 8, clumpSize);
-        BigEndian.setInt32(data, 12, totalBlock);
+        BigEndian.setInt32(data, 8, (int) clumpSize);
+        BigEndian.setInt32(data, 12, (int) totalBlock);
         for (int i = 0; i < extents.length; i++) {
             extents[i].write(data, EXTENT_OFFSET + (i * ExtentDescriptor.EXTENT_DESCRIPTOR_LENGTH));
         }
@@ -134,11 +133,11 @@ public class HfsPlusForkData {
         return totalSize;
     }
 
-    public int getClumpSize() {
+    public long getClumpSize() {
         return clumpSize;
     }
 
-    public int getTotalBlocks() {
+    public long getTotalBlocks() {
         return totalBlock;
     }
 
@@ -175,8 +174,8 @@ public class HfsPlusForkData {
      * Read a block of data
      *
      * @param fileSystem the associated file system.
-     * @param offset     the offset to read from.
-     * @param buffer     the buffer to read into.
+     * @param offset the offset to read from.
+     * @param buffer the buffer to read into.
      * @throws java.io.IOException if an error occurs.
      */
     public void read(HfsPlusFileSystem fileSystem, long offset, ByteBuffer buffer) throws IOException {
@@ -221,6 +220,7 @@ public class HfsPlusForkData {
     }
 
     /**
+     *
      * @param index
      * @param desc
      */
@@ -228,9 +228,7 @@ public class HfsPlusForkData {
         extents[index] = desc;
     }
 
-
     public ExtentDescriptor[] getExtents() {
         return extents;
     }
-
 }
